+++ src/IconFactory/IconFactory.tsx
@@ -13,6 +13,7 @@ import ChevronLeft from '../../assets/icons/system/chevron-left.svg'
 import ChevronRight from '../../assets/icons/system/chevron-right.svg'
 import ChevronUp from '../../assets/icons/system/chevron-up.svg'
 import Close from '../../assets/icons/system/close.svg'
+import CloudUpload from '../../assets/icons/system/cloud-upload.svg'
 import Data from '../../assets/icons/system/data.svg'
 import Download from '../../assets/icons/system/download.svg'
 import Edit from '../../assets/icons/system/edit.svg'
@@ -73,6 +74,7 @@ export type IconName =
   | 'add-product'
   | 'long-arrow-right'
   | 'download'
+  | 'cloud-upload'
 
 // Map icon names to their SVG components
 const iconMap: Record<IconName, React.FC<React.SVGProps<SVGSVGElement>>> = {
@@ -110,6 +112,7 @@ const iconMap: Record<IconName, React.FC<React.SVGProps<SVGSVGElement>>> = {
   'add-product': AddProductIcon,
   'long-arrow-right': LongArrowRight,
   download: Download,
+  'cloud-upload': CloudUpload,
 }
 
 export interface IconFactoryProps extends React.SVGProps<SVGSVGElement> {