+++ src/index.ts
@@ -18,6 +18,7 @@ export { default as Tooltip } from './Tooltip'
 export { default as Typography } from './Typography'
 export { default as IconFactory } from './IconFactory'
 export { default as Dropdown } from './Dropdown'
+export { default as FileUpload } from './FileUpload'
 
 // Types
 export type { IconName } from './IconFactory'